# optimize Cmp