```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/float16
        │  .old.txt   │              .new.txt              │
        │   sec/op    │   sec/op     vs base               │
Cmp-10    7.391n ± 1%   6.875n ± 1%  -6.99% (p=0.000 n=10)
Cmp2-10   11.41n ± 1%   11.50n ± 1%       ~ (p=0.077 n=10)
geomean   9.186n        8.894n       -3.18%
```